requirement.txt
arabic-reshaper==2.0.14
backports.functools-lru-cache==1.5
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
future==0.16.0
hazm==0.5.2
idna==2.6
kiwisolver==1.0.1
libwapiti==0.2.1
matplotlib==2.2.2
nltk==3.0.5
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
persian-wordcloud==1.3.1
Pillow==5.0.0
pkg-resources==0.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-bidi==0.4.0
python-dateutil==2.7.0
python-twitter==3.4.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
subprocess32==3.2.7
tweepy==3.6.0
urllib3==1.22
wordcloud==1.4.1
